refactor(maestro): keep model navigation helpers focused
