+++ crates/vize_maestro/src/ide/hover/corsa.rs
@@ -324,7 +324,7 @@ impl HoverService {
 }
 
 fn authored_hover_token_range(ctx: &IdeContext<'_>) -> Option<Range> {
-    if let Some((start, end)) = v_model_argument_token_span(&ctx.content, ctx.offset) {
+    if let Some((start, end)) = super::v_model::argument_token_span(&ctx.content, ctx.offset) {
         let (start_line, start_character) = crate::ide::offset_to_position(&ctx.content, start);
         let (end_line, end_character) = crate::ide::offset_to_position(&ctx.content, end);
         return Some(Range::new(
@@ -341,24 +341,4 @@ fn authored_hover_token_range(ctx: &IdeContext<'_>) -> Option<Range> {
         Position::new(start_line, start_character),
         Position::new(end_line, end_character),
     ))
-}
-
-fn v_model_argument_token_span(content: &str, offset: usize) -> Option<(usize, usize)> {
-    let (token_start, token_end) =
-        crate::ide::token_span_at_offset(content, offset, HoverService::is_word_char)?;
-    let token = content.get(token_start..token_end)?;
-    let argument = token.strip_prefix("v-model:")?;
-    if argument.starts_with('[') {
-        return None;
-    }
-
-    let argument_len = argument
-        .split_once('.')
-        .map_or(argument.len(), |(name, _)| name.len());
-    if argument_len == 0 {
-        return None;
-    }
-
-    let start = token_start + "v-model:".len();
-    Some((start, start + argument_len))
 }
